refactor: rewrite client code in Lua
TS is over-complicated and provides little benefit to me, especially when both TS and JS need to be bundled just to utilise in the V8 runtime.

From testing this should be a 1:1 rewrite (make an issue if something is broken).

WIP support for esx_skin/skinchanger replacement.

I intend to, eventually, rewrite any possible improvements in TS to PR into the original resource (so long as it remains in-scope).